requirements.txt
blessed==1.17.6
click==7.1.2
Flask==1.1.2
#Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
inquirer==2.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysql-connector-python==8.0.11
#mysqlclient==2.0.3
protobuf==3.14.0
python-editor==1.0.4
readchar==2.0.1
six==1.15.0
SQLAlchemy==1.3.22
wcwidth==0.2.5
Werkzeug==1.0.1
